backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
google-generativeai>=0.8.3
gspread==6.0.2
google-auth==2.28.1
google-auth-oauthlib==1.2.0
pandas==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.6.1
